requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
pycryptodome==3.21.0
itsdangerous==2.2.0
